+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 	- `sortBy` -  `'page'` (default) or `'title'`
 	- `sortLocale` - custom locale to sort by property `sortBy`
 - Added property `decorationThickness` for `text` to set width of the decoration line
+- Added inherited/extends styles, use property `extends` in style with style name or array of string with style names
 - Fixed margin override with 0 value
 - Fixed margin override from multiple styles
 - Fixed svg-to-pdfkit - TypeError: t.classList.contains is not a function